+++ lib/async/service/supervisor/envoy/supervised.rb
@@ -3,7 +3,6 @@
 # Released under the MIT License.
 # Copyright, 2026, by Samuel Williams.
 
-require "async/http/protocol/http2"
 require "async/service/supervisor/supervised"
 
 module Async
@@ -28,20 +27,9 @@ module Async
 						{
 							name: listener.name,
 							scheme: listener.scheme,
-							protocol: envoy_protocol(listener.protocol),
+							protocols: listener.protocols,
 							addresses: listener.addresses.filter_map{|address| envoy_address(address)},
 						}
-					end
-					
-					# Convert an application protocol implementation to an Envoy upstream protocol.
-					# @parameter protocol [Object] The application protocol implementation.
-					# @returns [Symbol] The Envoy upstream protocol.
-					def envoy_protocol(protocol)
-						if protocol.equal?(Async::HTTP::Protocol::HTTP2)
-							:http2
-						else
-							:http1
-						end
 					end
 					
 					# Convert a concrete bound address to serializable supervisor state.